+++ developer/src/kmcmplib/src/crc32.h
