+++ developer/src/kmcmplib/src/pch.h
@@ -7,7 +7,7 @@
 #include <km_u16.h>
 #include <kmcompx.h>
 
-#include "../../../../common/windows/cpp/include/crc32.h"
+#include "crc32.h"
 
 #include <ctype.h>
 #include <string.h>
