+++ developer/src/kmdecomp/savekeyboard.cpp
@@ -33,7 +33,7 @@
 
 #include "../../../common/windows/cpp/include/legacy_kmx_file.h"
 #include "../../../common/windows/cpp/include/legacy_kmx_memory.h"
-#include "../../../common/windows/cpp/include/vkeys.h"
+#include "../../../common/include/vkeys.h"
 
 #define SSN__PREFIX		L"&"
 
@@ -158,13 +158,13 @@ PWCHAR flagstr(int flag)
 	return buf;
 }
 
-PCWCHAR GetVKeyName(LPKEY key)  // I3438
+const char* GetVKeyName(LPKEY key)  // I3438
 {
-  static WCHAR buf[100];
+  static char buf[100];
   if(key->Key <= VK__MAX)
     return VKeyNames[key->Key];
 
-  wsprintfW(buf, L"%d", key->Key - 256);  //TODO: Support getting the key name from the VK Dictionary
+  wsprintfA(buf, "%d", key->Key - 256);  //TODO: Support getting the key name from the VK Dictionary
   return buf;
 }
 
@@ -178,7 +178,7 @@ PWCHAR KeyString(LPKEY key)
 			wsprintfW(buf, L"[%s%c%c%c] ", flagstr(key->ShiftFlags), key->Key == L'"' ? L'\'' : L'"',
 				key->Key, key->Key == L'"' ? L'\'' : L'"');
 		else
-			wsprintfW(buf, L"[%s%s] ", flagstr(key->ShiftFlags), GetVKeyName(key));////; //, VKeyNames[key->Key]);  // I3438
+			wsprintfW(buf, L"[%s%hs] ", flagstr(key->ShiftFlags), GetVKeyName(key));  // I3438
 	}
 	else
 	{
@@ -261,7 +261,7 @@ PWCHAR ExtString(PWCHAR str)
             //TODO: Get extended key value
             wsprintfW(p, L"[%s %d] ", flagstr(*str), *(str+1)-VK__MAX-1);
           else
-					  wsprintfW(p, L"[%s%s] ", flagstr(*str), VKeyNames[*(str+1)]);
+					  wsprintfW(p, L"[%s%hs] ", flagstr(*str), VKeyNames[*(str+1)]);
         }
 				str+=2; // skip UC_SENTINEL_EXTENDEDEND
 				p = wcschr(p, 0);
