+++ windows/src/engine/keyman32/keymanengine.h
@@ -257,8 +257,6 @@ void keybd_shift(LPINPUT pInputs, int* n, BOOL isReset, LPBYTE const kbd);
 #include "..\..\..\..\common\windows\cpp\include\keynames.h"
 #include "..\..\..\include\kmtip_guids.h"
 
-#include "..\..\..\..\common\windows\cpp\include\crc32.h"
-
 #include "k32_tsf.h"
 #include "k32_visualkeyboardinterface.h"
 #include "LowLevelHookWatchDog.h"
