+++ windows/src/engine/tsysinfox64/version.rc
@@ -1,4 +1,4 @@
-#include "../../../../common/windows/cpp/include/keymanversion.h"
+#include "../../../../common/include/keymanversion.h"
 
 1 VERSIONINFO
   FILEVERSION		KV_FILEVERSION
