+++ windows/src/test/manual-tests/test_i3619/i3619tip/i3619tip/i3619tip/version.rc
@@ -1,4 +1,4 @@
-#include "../../../../../../../../common/windows/cpp/include/keymanversion.h"
+#include "../../../../../../../../common/include/keymanversion.h"
 
 1 VERSIONINFO
   FILEVERSION		KV_FILEVERSION
